refactor: extract auth domain core
