+++ src/runtime/mod.rs
@@ -3351,11 +3351,9 @@ mod runtime_tests {
 #[cfg(test)]
 mod auth_tests {
     use super::*;
+    use crate::domain::auth::{AuthProviderEntry, OAuthFlowKind, OAuthStatus};
     use crate::handlers::*;
-    use crate::protocol::{
-        AuthMethod, AuthProviderEntry, AuthProvidersData, ClientMsg, OAuthFlowData, OAuthFlowKind,
-        OAuthResultData, OAuthStatus,
-    };
+    use crate::protocol::{ClientMsg, OAuthFlowData, OAuthResultData};
     use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 
     fn key(code: KeyCode) -> KeyEvent {
@@ -3411,91 +3409,6 @@ mod auth_tests {
         app.auth_providers = providers;
         app.popup = app::Popup::ProviderAuth;
         app
-    }
-
-    // ── Protocol type tests ───────────────────────────────────────────────────
-
-    #[test]
-    fn auth_provider_entry_effective_auth_oauth_connected() {
-        let mut p = make_provider("OpenAI");
-        p.oauth_status = Some(OAuthStatus::Connected);
-        assert_eq!(p.effective_auth(), Some(AuthMethod::OAuth));
-        assert_eq!(p.auth_badge_label(), "OAuth");
-        assert!(p.is_auth_active());
-    }
-
-    #[test]
-    fn auth_provider_entry_effective_auth_api_key_stored() {
-        let mut p = make_api_key_only("Groq");
-        p.has_stored_api_key = true;
-        assert_eq!(p.effective_auth(), Some(AuthMethod::ApiKey));
-        assert_eq!(p.auth_badge_label(), "API Key");
-        assert!(p.is_auth_active());
-    }
-
-    #[test]
-    fn auth_provider_entry_effective_auth_env_var() {
-        let mut p = make_api_key_only("DeepSeek");
-        p.has_env_api_key = true;
-        assert_eq!(p.effective_auth(), Some(AuthMethod::EnvVar));
-        assert_eq!(p.auth_badge_label(), "Env");
-    }
-
-    #[test]
-    fn auth_provider_entry_not_configured() {
-        let p = make_provider("OpenAI");
-        assert_eq!(p.effective_auth(), None);
-        assert_eq!(p.auth_badge_label(), "Not configured");
-        assert!(!p.is_auth_active());
-    }
-
-    #[test]
-    fn auth_provider_entry_expired_badge() {
-        let mut p = make_oauth_only("Codex");
-        p.oauth_status = Some(OAuthStatus::Expired);
-        assert_eq!(p.auth_badge_label(), "Expired");
-    }
-
-    #[test]
-    fn auth_provider_entry_unconfigurable() {
-        let p = AuthProviderEntry {
-            provider: "codex".into(),
-            display_name: "Codex".into(),
-            oauth_status: None,
-            has_stored_api_key: false,
-            has_env_api_key: false,
-            env_var_name: None,
-            supports_oauth: false,
-            preferred_method: None,
-        };
-        assert!(p.is_unconfigurable());
-        assert_eq!(p.auth_badge_label(), "OAuth required");
-    }
-
-    #[test]
-    fn auth_provider_entry_preferred_method_overrides_default_order() {
-        let mut p = make_provider("OpenAI");
-        p.oauth_status = Some(OAuthStatus::Connected);
-        p.has_stored_api_key = true;
-        p.preferred_method = Some(AuthMethod::ApiKey);
-        // With preference ApiKey, should pick that over OAuth even though both are available
-        assert_eq!(p.effective_auth(), Some(AuthMethod::ApiKey));
-    }
-
-    #[test]
-    fn auth_provider_classification_helpers() {
-        let multi = make_provider("OpenAI");
-        assert!(multi.has_multiple_auth_methods());
-        assert!(!multi.is_oauth_only());
-        assert!(!multi.is_api_key_only());
-
-        let oauth_only = make_oauth_only("Codex");
-        assert!(oauth_only.is_oauth_only());
-        assert!(!oauth_only.has_multiple_auth_methods());
-
-        let api_only = make_api_key_only("Groq");
-        assert!(api_only.is_api_key_only());
-        assert!(!api_only.has_multiple_auth_methods());
     }
 
     // ── App state tests ───────────────────────────────────────────────────────
@@ -4002,135 +3915,5 @@ mod auth_tests {
 
         let msg = rx.try_recv().expect("message sent");
         assert!(matches!(msg, ClientMsg::ListAuthProviders));
-    }
-
-    // ── ClientMsg serialization tests ─────────────────────────────────────────
-
-    #[test]
-    fn client_msg_list_auth_providers_serializes() {
-        let msg = ClientMsg::ListAuthProviders;
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("list_auth_providers"));
-    }
-
-    #[test]
-    fn client_msg_set_api_token_serializes() {
-        let msg = ClientMsg::SetApiToken {
-            provider: "openai".into(),
-            api_key: "sk-123".into(),
-        };
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("set_api_token"));
-        assert!(json.contains("openai"));
-        assert!(json.contains("sk-123"));
-    }
-
-    #[test]
-    fn client_msg_start_oauth_login_serializes() {
-        let msg = ClientMsg::StartOAuthLogin {
-            provider: "codex".into(),
-        };
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("start_oauth_login"));
-        assert!(json.contains("codex"));
-    }
-
-    #[test]
-    fn client_msg_complete_oauth_login_serializes() {
-        let msg = ClientMsg::CompleteOAuthLogin {
-            flow_id: "f1".into(),
-            response: "code123".into(),
-        };
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("complete_oauth_login"));
-        assert!(json.contains("f1"));
-    }
-
-    #[test]
-    fn client_msg_set_auth_method_serializes() {
-        let msg = ClientMsg::SetAuthMethod {
-            provider: "openai".into(),
-            method: AuthMethod::ApiKey,
-        };
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("set_auth_method"));
-        assert!(json.contains("api_key"));
-    }
-
-    // ── Deserialization round-trip tests (server → TUI) ───────────────────────
-
-    #[test]
-    fn auth_method_oauth_serde_matches_server_wire_format() {
-        // Server explicitly renames OAuth to "oauth" (not "o_auth")
-        let json = serde_json::json!("oauth");
-        let method: AuthMethod = serde_json::from_value(json).unwrap();
-        assert_eq!(method, AuthMethod::OAuth);
-
-        // Round-trip: our serialization must also produce "oauth"
-        let serialized = serde_json::to_string(&AuthMethod::OAuth).unwrap();
-        assert_eq!(serialized, "\"oauth\"");
-    }
-
-    #[test]
-    fn auth_provider_entry_deserializes_with_oauth_preferred_method() {
-        // Exact JSON shape the server sends for a provider with OAuth preference
-        let json = serde_json::json!({
-            "provider": "anthropic",
-            "display_name": "Anthropic",
-            "oauth_status": "connected",
-            "has_stored_api_key": true,
-            "has_env_api_key": false,
-            "env_var_name": "ANTHROPIC_API_KEY",
-            "supports_oauth": true,
-            "preferred_method": "oauth"
-        });
-        let entry: AuthProviderEntry = serde_json::from_value(json).unwrap();
-        assert_eq!(entry.provider, "anthropic");
-        assert_eq!(entry.oauth_status, Some(OAuthStatus::Connected));
-        assert_eq!(entry.preferred_method, Some(AuthMethod::OAuth));
-    }
-
-    #[test]
-    fn auth_providers_data_deserializes_mixed_providers() {
-        // Full server response shape
-        let json = serde_json::json!({
-            "providers": [
-                {
-                    "provider": "openai",
-                    "display_name": "OpenAI",
-                    "oauth_status": "not_authenticated",
-                    "has_stored_api_key": false,
-                    "has_env_api_key": true,
-                    "env_var_name": "OPENAI_API_KEY",
-                    "supports_oauth": true,
-                    "preferred_method": "oauth"
-                },
-                {
-                    "provider": "groq",
-                    "display_name": "Groq",
-                    "oauth_status": null,
-                    "has_stored_api_key": true,
-                    "has_env_api_key": false,
-                    "env_var_name": "GROQ_API_KEY",
-                    "supports_oauth": false,
-                    "preferred_method": "api_key"
-                },
-                {
-                    "provider": "local",
-                    "display_name": "Local",
-                    "oauth_status": null,
-                    "has_stored_api_key": false,
-                    "has_env_api_key": false,
-                    "env_var_name": null,
-                    "supports_oauth": false,
-                    "preferred_method": null
-                }
-            ]
-        });
-        let data: AuthProvidersData = serde_json::from_value(json).unwrap();
-        assert_eq!(data.providers.len(), 3);
-        assert_eq!(data.providers[0].preferred_method, Some(AuthMethod::OAuth));
-        assert_eq!(data.providers[1].preferred_method, Some(AuthMethod::ApiKey));
-        assert_eq!(data.providers[2].preferred_method, None);
     }
 }
